Move DiskDataCacheProvider to its own file and add write test
