+++ Common/Data/LeanDataWriter.cs
@@ -327,92 +327,4 @@ namespace QuantConnect.Data
         }
 
     }
-
-    /// <summary>
-    /// Simple data cache provider, writes and reads directly from disk
-    /// Used as default for <see cref="LeanDataWriter"/>
-    /// </summary>
-    public class DiskDataCacheProvider : IDataCacheProvider
-    { 
-        /// <summary>
-        /// Property indicating the data is temporary in nature and should not be cached.
-        /// </summary>
-        public bool IsDataEphemeral { get; }
-
-        /// <summary>
-        /// Simple data cache provider, writes and reads directly from disk
-        /// Used as default for <see cref="LeanDataWriter"/>
-        /// </summary>
-        public DiskDataCacheProvider()
-        {
-            IsDataEphemeral = false;
-        }
-
-        /// <summary>
-        /// Fetch data from the cache
-        /// </summary>
-        /// <param name="key">A string representing the key of the cached data</param>
-        /// <returns>An <see cref="Stream"/> of the cached data</returns>
-        public Stream Fetch(string key)
-        {
-            LeanData.ParseKey(key, out var filePath, out var entryName);
-
-            if (!File.Exists(filePath))
-            {
-                return null;
-            }
-
-            try
-            {
-                using (var zip = ZipFile.Read(filePath))
-                {
-                    ZipEntry entry;
-                    if (entryName.IsNullOrEmpty())
-                    {
-                        // Return the first entry
-                        entry = zip[0];
-                    }
-                    else
-                    {
-                        // Attempt to find our specific entry
-                        if (!zip.ContainsEntry(entryName))
-                        {
-                            return null;
-                        }
-
-                        entry = zip[entryName];
-                    }
-
-                    // Extract our entry and return it
-                    var stream = new MemoryStream();
-                    entry.Extract(stream);
-                    return stream;
-                }
-            }
-            catch (ZipException exception)
-            {
-                Log.Error("DiskDataCacheProvider.Fetch(): Corrupt file: " + key + " Error: " + exception);
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Store the data in the cache. Not implemented in this instance of the IDataCacheProvider
-        /// </summary>
-        /// <param name="key">The source of the data, used as a key to retrieve data in the cache</param>
-        /// <param name="data">The data as a byte array</param>
-        public void Store(string key, byte[] data)
-        {
-            LeanData.ParseKey(key, out var filePath, out var entryName);
-            Compression.ZipCreateAppendData(filePath, entryName, data);
-        }
-
-        /// <summary>
-        /// Dispose for this class
-        /// </summary>
-        public void Dispose()
-        {
-            //NOP
-        }
-    }
 }
